#18 Splited project at core and song bot. Many refactoring done
