--- src/main/java/com/example/telegrambotspring/controllers/MainController.java
@@ -1,12 +1,0 @@
-package com.example.telegrambotspring.controllers;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.RequestMapping;
-
-@Controller
-public class MainController {
-	@RequestMapping("/")
-	public String index() {
-		return "index.html";
-	}
-}
