--- src/main/java/com/example/telegrambotspring/repositories/SongsRepositoryCustom.java
@@ -1,7 +1,0 @@
-package com.example.telegrambotspring.repositories;
-
-import com.example.telegrambotspring.entities.SongCouplet;
-
-public interface SongsRepositoryCustom<T extends SongCouplet> {
-	T updateOrInsert(T object);
-}
