--- src/main/resources/application.properties
@@ -1,10 +1,0 @@
-logging.level.com.example.telegrambotspring=DEBUG
-server.port=80
-spring.data.mongodb.database=songs
-telegram.bot.textsDirectory=texts
-telegram.bot.longPoolingTimeout=30
-telegram.bot.token=
-spring.data.mongodb.uri=mongodb+srv://imAdmin:<password>@mycluster-az7yk.mongodb.net/songs?retryWrites=true&w=majority
-app.getting-updates-latency=5000
-app.send-message-latency.group=5000
-app.send-message-latency.direct=3000
