--- src/main/resources/static/delete.html
@@ -1,11 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Delete</title>
-</head>
-<body>
-Javascript should be enabled!
-<script src="scripts/script.js"></script>
-</body>
-</html>
