--- src/main/resources/static/edit.html
@@ -1,17 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Edit song</title>
-    <link href="styles/styles.css" media="all" rel="stylesheet" type="text/css">
-</head>
-<body>
-<form id="editSongForm" onsubmit="updateSong()">
-    <input name="artist" placeholder="Artist" type="text">
-    <input name="song" placeholder="Song" type="text">
-    <textarea name="text" placeholder="Text" rows="16"></textarea>
-    <button>Submit</button>
-    <script src="scripts/script.js"></script>
-</form>
-</body>
-</html>
