--- src/main/resources/static/index.html
@@ -1,15 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Hello</title>
-    <link href="styles/styles.css" media="all" rel="stylesheet" type="text/css">
-</head>
-
-<body>
-<div id="menu-container">
-    <a class="menu-item" href="list.html">List songs</a>
-    <a class="menu-item" href="add.html">Add song</a>
-</div>
-</body>
-</html>
