--- src/main/resources/static/list.html
@@ -1,15 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>List</title>
-    <link href="styles/styles.css" media="all" rel="stylesheet" type="text/css">
-</head>
-<body>
-<ul id="list-container">
-    Javascript should be enabled!
-    <script src="scripts/utils.js"></script>
-    <script src="scripts/script.js"></script>
-</ul>
-</body>
-</html>
