--- src/test/java/com/example/telegrambotspring/TelegramBotSpringApplicationTests.java
@@ -1,13 +1,0 @@
-package com.example.telegrambotspring;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class TelegramBotSpringApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
